requirements.txt
python-telegram-bot==20.7
httpx
requests
beautifulsoup4